Major refactor of code structure. Move classes to relevant packages. Some code commented out.
