--- src/main/java/com/mindlinksoft/recruitment/mychat/ConversationReader.java
@@ -1,5 +1,0 @@
-package com.mindlinksoft.recruitment.mychat;
-
-public class ConversationReader {
-    
-}
