--- src/main/java/com/mindlinksoft/recruitment/mychat/ConversationReaderService.java
@@ -1,5 +1,0 @@
-package com.mindlinksoft.recruitment.mychat;
-
-public interface ConversationReaderService {
-    
-}
