--- src/main/java/com/mindlinksoft/recruitment/mychat/ConversationWriter.java
@@ -1,5 +1,0 @@
-package com.mindlinksoft.recruitment.mychat;
-
-public class ConversationWriter {
-    
-}
