--- src/main/java/com/mindlinksoft/recruitment/mychat/ConversationWriterService.java
@@ -1,5 +1,0 @@
-package com.mindlinksoft.recruitment.mychat;
-
-public interface ConversationWriterService {
-    
-}
