+++ src/main/java/com/mindlinksoft/recruitment/mychat/exporter/datastructure/Conversation.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.mychat.exporter.datastructure;
 
 import java.util.List;
 import java.util.Map;
