+++ src/main/java/com/mindlinksoft/recruitment/mychat/exporter/datastructure/Message.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.mychat.exporter.datastructure;
 
 import java.time.Instant;
 import java.util.Map;
