+++ src/main/java/com/mindlinksoft/recruitment/mychat/exporter/reader/ConversationReader.java
@@ -1,0 +1,5 @@
+package com.mindlinksoft.recruitment.mychat.exporter.reader;
+
+public class ConversationReader {
+    
+}
