+++ src/main/java/com/mindlinksoft/recruitment/mychat/exporter/reader/ConversationReaderService.java
@@ -1,0 +1,5 @@
+package com.mindlinksoft.recruitment.mychat.exporter.reader;
+
+public interface ConversationReaderService {
+    
+}
