+++ src/main/java/com/mindlinksoft/recruitment/mychat/exporter/writer/ConversationWriter.java
@@ -1,0 +1,5 @@
+package com.mindlinksoft.recruitment.mychat.exporter.writer;
+
+public class ConversationWriter {
+    
+}
