+++ src/main/java/com/mindlinksoft/recruitment/mychat/exporter/writer/ConversationWriterService.java
@@ -1,0 +1,5 @@
+package com.mindlinksoft.recruitment.mychat.exporter.writer;
+
+public interface ConversationWriterService {
+    
+}
