+++ src/main/java/com/mindlinksoft/recruitment/mychat/main/CommandLineArgumentParser.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.mychat.main;
 
 /**
  * Represents a helper to parse command line arguments.
