+++ src/main/java/com/mindlinksoft/recruitment/mychat/main/ConversationExporterConfiguration.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.mychat.main;
 
 /**
  * Represents the configuration for the exporter.
