+++ src/main/java/com/mindlinksoft/recruitment/mychat/main/Main.java
@@ -1,0 +1,5 @@
+package com.mindlinksoft.recruitment.mychat.main;
+
+public class Main {
+    
+}
