--- src/test/java/com/mindlinksoft/recruitment/mychat/ConversationWriterTest.java
@@ -1,5 +1,0 @@
-package com.mindlinksoft.recruitment.mychat;
-
-public class ConversationWriterTest {
-    
-}
