--- src/test/java/com/mindlinksoft/recruitment/mychat/MainTests.java
@@ -1,5 +1,0 @@
-package com.mindlinksoft.recruitment.mychat;
-
-public class MainTests {
-    
-}
