+++ src/test/java/com/mindlinksoft/recruitment/mychat/exporter/reader/ConversationReaderTests.java
@@ -1,0 +1,5 @@
+package com.mindlinksoft.recruitment.mychat.exporter.reader;
+
+public class ConversationReaderTests {
+    
+}
