+++ src/test/java/com/mindlinksoft/recruitment/mychat/exporter/writer/ConversationWriterTest.java
@@ -1,0 +1,5 @@
+package com.mindlinksoft.recruitment.mychat.exporter.writer;
+
+public class ConversationWriterTest {
+    
+}
