+++ src/test/java/com/mindlinksoft/recruitment/mychat/main/ConversationExporterConfigurationTests.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.mychat.main;
 
 public class ConversationExporterConfigurationTests {
