+++ src/test/java/com/mindlinksoft/recruitment/mychat/main/MainTests.java
@@ -1,0 +1,5 @@
+package com.mindlinksoft.recruitment.mychat.main;
+
+public class MainTests {
+    
+}
